+++ src/main/java/com/bizcub/messageDisplayTime/Main.java
@@ -1,15 +1,11 @@
 package com.bizcub.messageDisplayTime;
 
-import com.bizcub.messageDisplayTime.config.ModConfig;
+import net.minecraft.client.OptionInstance;
 
 public class Main {
     public static final String MOD_ID = /*$ mod_id*/ "message_display_time";
-
-    public static ModConfig getConfig() {
-        return ModConfig.CONFIG;
-    }
+    public static OptionInstance<Integer> messageDisplayTimeSlider;
 
     public static void init() {
-        getConfig();
     }
 }