+++ src/main/java/com/bizcub/messageDisplayTime/config/ModConfig.java
@@ -1,9 +1,0 @@
-package com.bizcub.messageDisplayTime.config;
-
-public interface ModConfig {
-    ModConfig CONFIG = Compat.isClothConfigLoaded() ? ModClothConfig.getInstance() : new ModConfig() { };
-
-    default int sec() {
-        return 5;
-    }
-}