+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-* 26.1 update
+* Message display time can now be adjusted in the chat settings
+* Removed dependency on Cloth Config
+* Mod icon updated
+* Updated to 26.2